backend/requirements.txt
fastapi
uvicorn
pymongo
pydantic
pydantic-settings
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
email-validator
aiosmtplib
